refactor: Change mainHiddenChapter component
