+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "jest": "^27.5.1",
+    "nanoid": "^3.3.1",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
