+++ src/common/utils/makeKey.js
@@ -1,0 +1,7 @@
+import { customAlphabet } from "nanoid";
+
+const makeKey = () => {
+  return customAlphabet("123456789abcdef", 10);
+};
+
+export default makeKey;
