+++ src/feature/chapter/hiddenChapter/MainHiddenChapter.js
@@ -2,10 +2,13 @@ import React, { useState } from "react";
 
 import styled from "styled-components";
 
+import Quiz from "../../../common/components/Quiz";
+import quiz from "../../../data/mainHiddenChapterQuiz.json";
 import Ticket from "./Ticket";
 
 const MainHiddenChapter = () => {
   const [isOpen, setIsOpen] = useState(false);
+  const [value, setValue] = useState("");
 
   return (
     <>
@@ -14,18 +17,7 @@ const MainHiddenChapter = () => {
           <img src="/assets/corridor.png" height="100%" alt="복도 이미지" />
         </ImgSection>
         <Ticket setIsOpen={setIsOpen} isOpen={isOpen} />
-
-        {isOpen && (
-          <QuizSection>
-            <div className="img-box">
-              <img src="/assets/file-folder.png" alt="폴더 이미지" />
-            </div>
-            <div className="text-box">
-              <h3></h3>
-              <div></div>
-            </div>
-          </QuizSection>
-        )}
+        {isOpen && <Quiz setValue={setValue} value={value} quiz={quiz} />}
       </Main>
     </>
   );
@@ -41,34 +33,6 @@ const Main = styled.main`
 const ImgSection = styled.section`
   width: 55%;
   height: 100vh;
-`;
-
-const QuizSection = styled.section`
-  position: relative;
-  width: 45%;
-  height: 100vh;
-
-  .img-box {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-
-    img {
-      width: 95%;
-      height: 93vh;
-    }
-  }
-
-  .text-box {
-    position: absolute;
-    top: 160px;
-    left: 120px;
-
-    h3 {
-      color: red;
-    }
-  }
 `;
 
 export default MainHiddenChapter;
