+++ src/feature/chapter/secondChapter/NewsPaper.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const NewsPaper = () => {
+  return <></>;
+};
+
+export default NewsPaper;
