+++ src/feature/chapter/secondChapter/Photo.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Photo = () => {
+  return <>포토이미지..</>;
+};
+
+export default Photo;
